+++ app/src/main/java/com/example/recipeplanner/MyContext.kt
@@ -2,7 +2,7 @@ package com.example.recipeplanner
 
 import android.content.Context
 import com.example.recipeplanner.ui.recipeBook.RecipeBookViewModelFactory
-import com.example.recipeplanner.ui.theme.createRecipe.CreateRecipeViewModelFactory
+import com.example.recipeplanner.ui.createRecipe.CreateRecipeViewModelFactory
 
 class MyContext(applicationContext: Context) {
     val recipeBookViewModelFactory = RecipeBookViewModelFactory(applicationContext)
