+++ app/src/main/java/com/example/recipeplanner/ui/createRecipe/CreateRecipeUIState.kt
@@ -1,4 +1,4 @@
-package com.example.recipeplanner.ui.theme.createRecipe
+package com.example.recipeplanner.ui.createRecipe
 
 import com.example.recipeplanner.data.Ingredient
